+++ docs/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add CISA CVSSIssuer type (OSIDB-4203)
 - Ignore modification of non-Red Hat issued CVSS scores through API
 
+### Changed
+- Adjust cve.org collector to correctly label CISA-ADP CVSS scores as CISA (OSIDB-4204)
+
 ### Fixed
 - Prevent running multiple Jira transition managers at the same time (OSIDB-4175)